Add calcDuration method and update time calculations
A new method `calcDuration(int distance)` was added to the `Car` class in `Car.h` and implemented in `Car.cpp`. This method calculates the duration required to cover a given distance based on the car's speed, using the `ceil` function to round up to the nearest whole number.

The `#pragma once` directive was added at the beginning of `Car.cpp` to ensure the file is included only once during compilation.

The `#include <cmath>` directive was added to `Car.cpp` to use the `ceil` function.

The `nextPickupTime` and `nextDropOffTime` calculations in the `assign` and `pickUp` methods were updated to use the new `calcDuration` method instead of directly performing the division.